Move assertions to it's correct extensions, enabling it's reuse by any `GalleryStoreSpects` implementation
